requirements.txt
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorama==0.4.5
cupshelpers==1.0
dbus-python==1.2.16
defusedxml==0.7.1
distlib==0.3.4
distro-info==1.0
dominate==2.7.0
et-xmlfile==1.1.0
filelock==3.7.1
Flask==2.1.3
Flask-Admin==1.6.0
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fuse-python==1.0.2
gpg===1.14.0-unknown
greenlet==1.1.2
httplib2==0.18.1
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
louis==3.16.0
MarkupSafe==2.1.1
numpy==1.23.1
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pdf2image==1.16.0
Pillow==9.2.0
platformdirs==2.5.2
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-ods==0.6.0
pyexcel-webio==0.1.4
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
PyGObject==3.38.0
pylibacl==0.6.0
pyodbc==4.0.34
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pytesseract==0.3.9
python-apt==2.2.1
python-dateutil==2.8.2
python-debian==0.1.39
python-debianbts==3.1.0
python-dotenv==0.20.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.39
texttable==1.6.4
tornado==6.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.14.1
visitor==0.1.3
Werkzeug==2.2.1
WTForms==3.0.1
xdg==5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
